Remove goal struct things compile now

+++ contracts/alarmClient/IAlarmClient.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.6.12;
 
 interface IAlarmClient {
-    function setAlarm(address pactAddress, uint64 endDateUtc) external;
+    function setAlarm(address pactAddress, uint endDateUtc) external;
     function fulfill(bytes32 requestId) external;
 }
